apt: Prevent installation of recommended packages
This is a bit of a split problem: we run an embedded Linux distribution and
don't need most of the packages that desktops have, but at the same time, we
don't want to break things. This allows us to err on the side of still remaining
functional, while causing some small breakages here and there.

Change-Id: I27ec72afb147ce14ed26f640427b303dbf403b6b