rootfs-cleanups: Migrate postinst steps into a runonce script
These scripts change settings in the system that little to do with a package,
and really have to do with the initial state of the installation. These steps
are usually taken by the debian-installer, but since we don't have one, we
really should take care of it in a runonce.

Change-Id: Ib0a772af9817d2ad045df4fe5f18fc81680b73e7