profile: Set the default LANG var if it is unset
Normally it would be a gaffe to set this anywhere by default, since ssh clients
are configured by default to send LANG along to the servers. Unfortunately, to
support "mdt shell" properly, we have to do this, as the current stable version
of paramiko does not provide a facility to set environment variables, and thus
LANG ends up unset.

Change-Id: Ifba3e5c10465759b41af553b225fa5a6ad654f5f